Repeat: Fixed #1, changed mvc-repeat from class to tag name
Major fix, had to redo most of it. Basically the way the repeat module was written only allowed the expression to be the only thing inside the mvc-repeat. This is a problem if you have multiple expressions or addittional html. So I had to redo allot of it.

Also changed mvc-repeat from a class name to a tag name